+++ tests/testthat/test_TunerSuccessiveHalving.R
@@ -1,11 +1,54 @@
-test_that("TunerSuccessiveHalving", {
-  skip_if_not_installed("mlr3learners")
-  skip_if_not_installed("xgboost")
-  library(mlr3learners) # nolint
+test_that("TunerAsha works", {
+  learner = lrn("classif.debug",
+    x  = to_tune(),
+    iter = to_tune(p_int(1, 16, tags = "budget"))
+  )
 
-  test_tuner_successive_halving(n = 16, eta = 2, n_dim = 1)
-  test_tuner_successive_halving(n = 16, eta = 2, n_dim = 2)
-  test_tuner_successive_halving(n = 16, eta = 2, lower_bound = 2, n_dim = 1)
-  test_tuner_successive_halving(n = 81, eta = 3, lower_bound = 1, upper_bound = 81, n_dim = 1)
+  test_tuner_successive_halving(n = 16, eta = 2, learner)
 })
 
+test_that("TunerAsha works with r_max > n", {
+  learner = lrn("classif.debug",
+    x  = to_tune(),
+    iter = to_tune(p_int(1, 17, tags = "budget"))
+  )
+
+  test_tuner_successive_halving(n = 16, eta = 2, learner)
+})
+
+test_that("TunerAsha works with r_max < n", {
+  learner = lrn("classif.debug",
+    x  = to_tune(),
+    iter = to_tune(p_int(1, 15, tags = "budget"))
+  )
+
+  test_tuner_successive_halving(n = 16, eta = 2, learner)
+})
+
+test_that("repeating successive halving works", {
+  learner = lrn("classif.debug",
+    x  = to_tune(),
+    iter = to_tune(p_int(1, 16, tags = "budget"))
+  )
+
+  instance = tune(
+    method = "successive_halving",
+    task = tsk("pima"),
+    learner = learner,
+    resampling = rsmp("cv", folds = 3),
+    measures = msr("classif.ce"),
+    repeats = TRUE,
+    term_evals = 62)
+
+  expect_equal(nrow(instance$archive$data), 62)
+
+  instance = tune(
+    method = "successive_halving",
+    task = tsk("pima"),
+    learner = learner,
+    resampling = rsmp("cv", folds = 3),
+    measures = msr("classif.ce"),
+    repeats = FALSE)
+
+  expect_equal(nrow(instance$archive$data), 31)
+})