+++ NEWS.md
@@ -1,7 +1,13 @@
+# mlr3hyperband 0.3.0
+
+* Implements parallel evaluation of stages.
+* Number of repetitions can be set with the `repeats` parameter.
+* Fixed maximization bug.
+
 # mlr3hyperband 0.2.0
 
-- Adds `OptimizerHyperband`.
-- Adds `OptimizerSuccessiveHalving`.
+* Adds `OptimizerHyperband`.
+* Adds `OptimizerSuccessiveHalving`.
 
 # mlr3hyperband 0.1.2